feat: map content and boundaries validation.

+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: lseabra- <lseabra-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/17 11:06:19 by lseabra-          #+#    #+#             */
-/*   Updated: 2026/03/26 16:48:06 by lseabra-         ###   ########.fr       */
+/*   Updated: 2026/03/27 17:08:15 by lseabra-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ typedef struct s_map
 {
 	char		grid[MAX_MAP_ROWS][MAX_LINE_SIZE];
 	t_spawn		player_spawn;
-	size_t		row_count;
+	int			row_count;
 }	t_map;
 
 typedef struct s_data
